+++ incapsula/client_incap_rule.go
@@ -44,7 +44,7 @@ func (c *Client) AddIncapRule(siteID string, rule *IncapRule) (*IncapRuleWithID,
 
 	// Post form to Incapsula
 	resp, err := c.httpClient.Post(
-		fmt.Sprintf("%s/sites/%s/rules?api_id=%s&api_key=%s", c.config.IncapRuleBaseURL, siteID, c.config.APIID, c.config.APIKey),
+		fmt.Sprintf("%s/sites/%s/rules?api_id=%s&api_key=%s", c.config.APIV2BaseURL, siteID, c.config.APIID, c.config.APIKey),
 		"application/json",
 		bytes.NewReader(ruleJSON))
 	if err != nil {
@@ -78,7 +78,7 @@ func (c *Client) ReadIncapRule(siteID string, ruleID int) (*IncapRuleWithID, int
 	log.Printf("[INFO] Getting Incapsula Incap Rule %d for Site ID %s\n", ruleID, siteID)
 
 	// Post form to Incapsula
-	resp, err := c.httpClient.Get(fmt.Sprintf("%s/sites/%s/rules/%d?api_id=%s&api_key=%s", c.config.IncapRuleBaseURL, siteID, ruleID, c.config.APIID, c.config.APIKey))
+	resp, err := c.httpClient.Get(fmt.Sprintf("%s/sites/%s/rules/%d?api_id=%s&api_key=%s", c.config.APIV2BaseURL, siteID, ruleID, c.config.APIID, c.config.APIKey))
 	if err != nil {
 		return nil, 0, fmt.Errorf("Error from Incapsula service when reading Incap Rule %d for Site ID %s: %s", ruleID, siteID, err)
 	}
@@ -117,7 +117,7 @@ func (c *Client) UpdateIncapRule(siteID string, ruleID int, rule *IncapRule) (*I
 	// Put request to Incapsula
 	req, err := http.NewRequest(
 		http.MethodPut,
-		fmt.Sprintf("%s/sites/%s/rules/%d?api_id=%s&api_key=%s", c.config.IncapRuleBaseURL, siteID, ruleID, c.config.APIID, c.config.APIKey),
+		fmt.Sprintf("%s/sites/%s/rules/%d?api_id=%s&api_key=%s", c.config.APIV2BaseURL, siteID, ruleID, c.config.APIID, c.config.APIKey),
 		bytes.NewReader(ruleJSON))
 	if err != nil {
 		return nil, fmt.Errorf("Error preparing HTTP PUT for updating Incap Rule %d for Site ID %s: %s", ruleID, siteID, err)
@@ -156,7 +156,7 @@ func (c *Client) DeleteIncapRule(siteID string, ruleID int) error {
 	// Delete request to Incapsula
 	req, err := http.NewRequest(
 		http.MethodDelete,
-		fmt.Sprintf("%s/sites/%s/rules/%d?api_id=%s&api_key=%s", c.config.IncapRuleBaseURL, siteID, ruleID, c.config.APIID, c.config.APIKey),
+		fmt.Sprintf("%s/sites/%s/rules/%d?api_id=%s&api_key=%s", c.config.APIV2BaseURL, siteID, ruleID, c.config.APIID, c.config.APIKey),
 		nil)
 	if err != nil {
 		return fmt.Errorf("Error preparing HTTP DELETE for deleting Incap Rule %d for Site ID %s: %s", ruleID, siteID, err)