+++ incapsula/resource_site.go
@@ -5,6 +5,7 @@ import (
 	"log"
 	"strconv"
 	"strings"
+	"time"
 
 	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
 )
@@ -62,11 +63,6 @@ func resourceSite() *schema.Resource {
 				Type:        schema.TypeString,
 				Optional:    true,
 			},
-			"log_level": &schema.Schema{
-				Description: "Available only for Enterprise Plan customers that purchased the Logs Integration SKU. Sets the log reporting level for the site. Options are full, security, none, and default.",
-				Type:        schema.TypeString,
-				Optional:    true,
-			},
 			"logs_account_id": &schema.Schema{
 				Description: "Available only for Enterprise Plan customers that purchased the Logs Integration SKU. Numeric identifier of the account that purchased the logs integration SKU and which collects the logs. If not specified, operation will be performed on the account identified by the authentication parameters.",
 				Type:        schema.TypeString,
@@ -112,6 +108,37 @@ func resourceSite() *schema.Resource {
 				Type:        schema.TypeString,
 				Optional:    true,
 			},
+			"data_storage_region": &schema.Schema{
+				Description: "The data region to use. Options are `APAC`, `AU`, `EU`, and `US`.",
+				Type:        schema.TypeString,
+				Computed:    true,
+				Optional:    true,
+			},
+			"hashing_enabled": &schema.Schema{
+				Description: "Specify if hashing (masking setting) should be enabled.",
+				Type:        schema.TypeBool,
+				Computed:    true,
+				Optional:    true,
+			},
+			"hash_salt": &schema.Schema{
+				Description: "Specify the hash salt (masking setting), required if hashing is enabled. Maximum length of 64 characters.",
+				Type:        schema.TypeString,
+				Computed:    true,
+				Optional:    true,
+				ValidateFunc: func(val interface{}, key string) (warns []string, errs []error) {
+					salt := val.(string)
+					if len(salt) > 64 {
+						errs = append(errs, fmt.Errorf("%q must be a max of 64 characters, got: %s", key, salt))
+					}
+					return
+				},
+			},
+			"log_level": &schema.Schema{
+				Description: "The log level. Options are `full`, `security`, and `none`. Defaults to `none`.",
+				Type:        schema.TypeString,
+				Default:     "none",
+				Optional:    true,
+			},
 
 			// Computed Attributes
 			"site_creation_date": &schema.Schema{
@@ -142,6 +169,11 @@ func resourceSite() *schema.Resource {
 					Type: schema.TypeString,
 				},
 			},
+			"domain_verification": &schema.Schema{
+				Description: "Domain verification (e.g. GlobalSign verification).",
+				Type:        schema.TypeString,
+				Computed:    true,
+			},
 		},
 	}
 }
@@ -170,19 +202,14 @@ func resourceSiteCreate(d *schema.ResourceData, m interface{}) error {
 	d.SetId(strconv.Itoa(siteAddResponse.SiteID))
 	log.Printf("[INFO] Created Incapsula site for domain: %s\n", domain)
 
-	// list of params from config that are specific to update after site creation
-	updateParams := [6]string{"active", "acceleration_level", "seal_location", "domain_redirect_to_full", "remove_ssl", "ignore_ssl"}
-	for i := 0; i < len(updateParams); i++ {
-		param := updateParams[i]
-		if d.Get(param) != "" {
-			log.Printf("[INFO] Updating Incapsula site param (%s) with value (%s) for site_id: %d\n", param, d.Get(param).(string), siteAddResponse.SiteID)
-			_, err := client.UpdateSite(strconv.Itoa(siteAddResponse.SiteID), param, d.Get(param).(string))
-			if err != nil {
-				log.Printf("[ERROR] Could not update Incapsula site param (%s) with value (%s) for site_id: %d %s\n", param, d.Get(param).(string), siteAddResponse.SiteID, err)
-				return err
-			}
-		}
-	}
+	// There may be a timing/race condition here
+	// Set an arbitrary period to sleep
+	time.Sleep(3 * time.Second)
+
+	updateAdditionalSiteProperties(client, d)
+	updateDataStorageRegion(client, d)
+	updateMaskingSettings(client, d)
+	updateLogLevel(client, d)
 
 	// Set the rest of the state from the resource read
 	return resourceSiteRead(d, m)
@@ -228,6 +255,33 @@ func resourceSiteRead(d *schema.ResourceData, m interface{}) error {
 	}
 	d.Set("dns_a_record_value", dnsARecordValues)
 
+	// Set the GlobalSign verification (may not exist)
+	if siteStatusResponse.Ssl.GeneratedCertificate.ValidationMethod == "dns" {
+		d.Set("domain_verification", siteStatusResponse.Ssl.GeneratedCertificate.ValidationData[0].SetDataTo[0])
+	}
+
+	// Get the log level for the site
+	if siteStatusResponse.LogLevel != "" {
+		d.Set("log_level", siteStatusResponse.LogLevel)
+	}
+
+	// Get the data storage region for the site
+	dataStorageRegionResponse, err := client.GetDataStorageRegion(d.Id())
+	if err != nil {
+		log.Printf("[ERROR] Could not read Incapsula site data storage region for domain: %s and site id: %d, %s\n", domain, siteID, err)
+		return err
+	}
+	d.Set("data_storage_region", dataStorageRegionResponse.Region)
+
+	// Get the masking settings for the site
+	maskingResponse, err := client.GetMaskingSettings(d.Id())
+	if err != nil {
+		log.Printf("[ERROR] Could not read Incapsula site masking settings for domain: %s and site id: %d, %s\n", domain, siteID, err)
+		return err
+	}
+	d.Set("hashing_enabled", maskingResponse.HashingEnabled)
+	d.Set("hash_salt", maskingResponse.HashSalt)
+
 	log.Printf("[INFO] Read Incapsula site for domain: %s\n", domain)
 
 	return nil
@@ -235,20 +289,11 @@ func resourceSiteRead(d *schema.ResourceData, m interface{}) error {
 
 func resourceSiteUpdate(d *schema.ResourceData, m interface{}) error {
 	client := m.(*Client)
-	siteID, _ := strconv.Atoi(d.Id())
 
-	updateParams := [6]string{"active", "acceleration_level", "seal_location", "domain_redirect_to_full", "remove_ssl", "ignore_ssl"}
-	for i := 0; i < len(updateParams); i++ {
-		param := updateParams[i]
-		if d.Get(param) != "" {
-			log.Printf("[INFO] Updating Incapsula site param (%s) with value (%s) for site_id: %d\n", param, d.Get(param).(string), siteID)
-			_, err := client.UpdateSite(strconv.Itoa(siteID), param, d.Get(param).(string))
-			if err != nil {
-				log.Printf("[ERROR] Could not update Incapsula site param (%s) with value (%s) for site_id: %d %s\n", param, d.Get(param).(string), siteID, err)
-				return err
-			}
-		}
-	}
+	updateAdditionalSiteProperties(client, d)
+	updateDataStorageRegion(client, d)
+	updateMaskingSettings(client, d)
+	updateLogLevel(client, d)
 
 	// Set the rest of the state from the resource read
 	return resourceSiteRead(d, m)
@@ -274,5 +319,59 @@ func resourceSiteDelete(d *schema.ResourceData, m interface{}) error {
 
 	log.Printf("[INFO] Deleted Incapsula site for domain: %s\n", domain)
 
+	return nil
+}
+
+func updateAdditionalSiteProperties(client *Client, d *schema.ResourceData) error {
+	updateParams := [7]string{"acceleration_level", "active", "approver", "domain_redirect_to_full", "domain_validation", "ignore_ssl", "remove_ssl"}
+	for i := 0; i < len(updateParams); i++ {
+		param := updateParams[i]
+		if d.HasChange(param) && d.Get(param) != "" {
+			log.Printf("[INFO] Updating Incapsula site param (%s) with value (%s) for site_id: %s\n", param, d.Get(param).(string), d.Id())
+			_, err := client.UpdateSite(d.Id(), param, d.Get(param).(string))
+			if err != nil {
+				log.Printf("[ERROR] Could not update Incapsula site param (%s) with value (%s) for site_id: %s %s\n", param, d.Get(param).(string), d.Id(), err)
+				return err
+			}
+		}
+	}
+	return nil
+}
+
+func updateDataStorageRegion(client *Client, d *schema.ResourceData) error {
+	if d.HasChange("data_storage_region") {
+		dataStorageRegion := d.Get("data_storage_region").(string)
+		_, err := client.UpdateDataStorageRegion(d.Id(), dataStorageRegion)
+		if err != nil {
+			log.Printf("[ERROR] Could not set Incapsula site data storage region with value (%s) for site_id: %s %s\n", dataStorageRegion, d.Id(), err)
+			return err
+		}
+	}
+	return nil
+}
+
+func updateMaskingSettings(client *Client, d *schema.ResourceData) error {
+	if d.HasChange("hashing_enabled") || d.HasChange("hash_salt") {
+		hashingEnabled := d.Get("hashing_enabled").(bool)
+		hashSalt := d.Get("hash_salt").(string)
+		maskingSettings := MaskingSettings{HashingEnabled: hashingEnabled, HashSalt: hashSalt}
+		err := client.UpdateMaskingSettings(d.Id(), &maskingSettings)
+		if err != nil {
+			log.Printf("[ERROR] Could not update Incapsula site masking settings for site_id: %s %s\n", d.Id(), err)
+			return err
+		}
+	}
+	return nil
+}
+
+func updateLogLevel(client *Client, d *schema.ResourceData) error {
+	if d.HasChange("log_level") {
+		logLevel := d.Get("log_level").(string)
+		err := client.UpdateLogLevel(d.Id(), logLevel)
+		if err != nil {
+			log.Printf("[ERROR] Could not update Incapsula site log level: %s for site_id: %s %s\n", logLevel, d.Id(), err)
+			return err
+		}
+	}
 	return nil
 }